# portless-daemon/api/openapi.yaml
openapi: 3.1.0
info:
  title: Portless local control API
  version: 10.7.0
  description: >-
    Loopback-only API shared by the Portless CLI and embedded UI. Projects are
    reusable application topologies. Environments are independently configured
    and runnable instances. Public routes use readable project, environment,
    service, and artifact names; private ownership keys are never exposed.
    Authenticated native clients may send Portless-Client-Kind as `cli` or
    `mcp`; the daemon records that fixed actor category on durable operations.
    Browser sessions cannot claim a native client kind.
servers:
  - url: http://portless.localhost/api/v1
security:
  - bearer: []
  - browserSession: []
paths:
...
  /environments/{projectName}/{environmentName}/traffic:
    parameters: *environmentParameters
    delete:
      operationId: clearTraffic
      description: Clears the live exchange buffer and derived traces without deleting durable recordings or resetting sequence numbers.
      responses:
        '200':
          description: Live traffic was cleared through the reported environment-local sequence.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrafficClearResponse' }
...
  /environments/{projectName}/{environmentName}/recordings:
    parameters: *environmentParameters
    get:
      operationId: listRecordings
      parameters:
        - { $ref: '#/components/parameters/Limit1000' }
      responses:
        '200':
          description: Environment recordings
          content:
            application/json:
              schema:
                type: object
                required: [recordings]
                properties:
                  recordings: { type: array, items: { $ref: '#/components/schemas/Recording' } }
    post:
      operationId: startRecording
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecordingInput' }
      responses:
        '201':
          description: Active bounded recording
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Recording' }
  /environments/{projectName}/{environmentName}/recordings/{recordingName}:
    parameters: &recordingParameters
      - { $ref: '#/components/parameters/ProjectName' }
      - { $ref: '#/components/parameters/EnvironmentName' }
      - { $ref: '#/components/parameters/RecordingName' }
    get:
      operationId: getRecording
      responses:
        '200':
          description: Recording metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Recording' }
    delete:
      operationId: deleteRecording
      responses:
        '204': { description: Completed recording deleted }
  /environments/{projectName}/{environmentName}/recordings/{recordingName}/stop:
    parameters: *recordingParameters
    post:
      operationId: stopRecording
      responses:
        '200':
          description: Completed recording
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Recording' }
...
  /environments/{projectName}/{environmentName}/faults:
    parameters: *environmentParameters
    get:
      operationId: listFaults
      parameters:
        - { $ref: '#/components/parameters/Limit1000' }
      responses:
        '200':
          description: Environment fault rules
          content:
            application/json:
              schema:
                type: object
                required: [faults]
                properties:
                  faults: { type: array, items: { $ref: '#/components/schemas/FaultRule' } }
    post:
      operationId: createFault
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FaultInput' }
      responses:
        '201':
          description: Enabled fault rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FaultRule' }
  /environments/{projectName}/{environmentName}/faults/{faultName}:
    parameters:
      - { $ref: '#/components/parameters/ProjectName' }
      - { $ref: '#/components/parameters/EnvironmentName' }
      - { $ref: '#/components/parameters/FaultName' }
    get:
      operationId: getFault
      responses:
        '200':
          description: Fault rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FaultRule' }
    delete:
      operationId: disableFault
      responses:
        '204': { description: Rule disabled; audit history retained }
  /environments/{projectName}/{environmentName}/faults/disable-all:
    parameters: *environmentParameters
    post:
      operationId: disableAllFaults
      responses:
        '200': { description: Active rules disabled atomically }
...
  /environments/{projectName}/{environmentName}/timeline:
    parameters: *environmentParameters
    get:
      operationId: timeline
      parameters:
        - { $ref: '#/components/parameters/Limit1000' }
      responses:
        '200':
          description: Newest-first durable environment timeline
          content:
            application/json:
              schema:
                type: object
                required: [timeline]
                properties:
                  timeline: { type: array, items: { $ref: '#/components/schemas/TimelineEvent' } }
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Mode-0600 installation secret used by the local CLI.
    browserSession:
      type: apiKey
      in: cookie
      name: portless_session
      description: SameSite=Strict browser session; mutations also require X-Portless-CSRF.
  parameters:
...
    ProjectName:
      in: path
      name: projectName
      required: true
      schema: { $ref: '#/components/schemas/ProjectName' }
    EnvironmentName:
      in: path
      name: environmentName
      required: true
      schema: { $ref: '#/components/schemas/EnvironmentName' }
...
    RecordingName:
      in: path
      name: recordingName
      required: true
      schema: { $ref: '#/components/schemas/ArtifactName' }
    FaultName:
      in: path
      name: faultName
      required: true
      schema: { $ref: '#/components/schemas/ArtifactName' }
...
    Limit1000:
      in: query
      name: limit
      description: Positive result bound; the endpoint applies its documented default when omitted.
      schema: { type: integer, minimum: 1, maximum: 1000 }
...
  schemas:
    ProjectName: &dnsName
      type: string
      pattern: '^[a-z][a-z0-9-]{0,62}$'
      examples: [billing]
...
    ArtifactName:
      type: string
      pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
      examples: [checkout-debug]
...
    TrafficClearResponse:
      type: object
      required: [cleared, throughSequence]
      properties:
        cleared: { type: integer, minimum: 0, description: Number of live exchanges removed. }
        throughSequence: { type: integer, format: int64, minimum: 0, description: Highest environment-local sequence covered by the clear. }
...
    RecordingInput:
      type: object
      required: [name]
      properties:
        name: { $ref: '#/components/schemas/ArtifactName' }
        source: { type: string }
        target: { type: string }
        captureBodies: { type: boolean, default: false, description: Persist bounded inspectable request and response bodies after secret-like headers are redacted. }
        maxEvents: { type: integer, default: 10000, maximum: 100000 }
        maxBodyBytes: { type: integer, default: 65536, maximum: 1048576 }
        expiresAt: { type: string, format: date-time }
    Recording:
      allOf:
        - { $ref: '#/components/schemas/RecordingInput' }
        - type: object
          required: [project, environment, status, startedAt, eventCount]
          properties:
            project: { $ref: '#/components/schemas/ProjectName' }
            environment: { $ref: '#/components/schemas/EnvironmentName' }
            status: { type: string, enum: [active, completed, failed] }
            startedAt: { type: string, format: date-time }
            completedAt: { type: string, format: date-time }
            eventCount: { type: integer, format: int64, minimum: 0 }
...
    FaultInput:
      type: object
      required: [name, source, target]
      properties:
        name: { $ref: '#/components/schemas/ArtifactName' }
        source: { type: string }
        target: { type: string }
        method: { type: string }
        path: { type: string }
        probability: { type: number, minimum: 0, maximum: 1, default: 1 }
        latencyMs: { type: integer, minimum: 0 }
        jitterMs: { type: integer, minimum: 0 }
        statusCode: { type: integer, minimum: 400, maximum: 599 }
        abort: { type: boolean }
        expiresAt: { type: string, format: date-time }
    FaultRule:
      allOf:
        - { $ref: '#/components/schemas/FaultInput' }
        - type: object
          required: [project, environment, enabled, createdAt, matchCount, revision, scopeSummary]
          properties:
            project: { $ref: '#/components/schemas/ProjectName' }
            environment: { $ref: '#/components/schemas/EnvironmentName' }
            enabled: { type: boolean }
            createdAt: { type: string, format: date-time }
            matchCount: { type: integer, format: int64, minimum: 0 }
            revision: { type: integer, format: int64, minimum: 1 }
            scopeSummary: { type: string }
    TimelineEvent:
      type: object
      required: [project, environment, sequence, timestamp, actor, type, severity, summary]
      properties:
        project: { $ref: '#/components/schemas/ProjectName' }
        environment: { $ref: '#/components/schemas/EnvironmentName' }
        sequence: { type: integer, format: int64, minimum: 1 }
        timestamp: { type: string, format: date-time }
        actor: { type: string }
        type: { type: string }
        subject: { type: string }
        severity: { type: string, enum: [info, warning, error] }
        summary: { type: string }
        details: { type: object, additionalProperties: true }
...